requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
chardet==4.0.0
decorator==5.1.0
discord.py==1.7.3
idna==3.3
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
multidict==5.2.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
python-dotenv==0.19.1
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
yarl==1.7.2
